requirements.txt
streamlit
audio-recorder-streamlit
gTTS
python-dotenv
SpeechRecognition
requests